Apply temporary patch for large DBF files. (#393)
git-svn-id: http://svn.osgeo.org/postgis/branches/1.5@5787 b70326c6-7e19-0410-871a-916f4a2858ee

+++ loader/dbfopen.c
@@ -741,7 +741,7 @@ DBFAddField(DBFHandle psDBF, const char * pszFieldName,
 
 int DBFReadSetup(DBFHandle psDBF, int hEntity)
 {
-	int	       	nRecordOffset;
+	long	       	nRecordOffset;
 
 	/* -------------------------------------------------------------------- */
 	/*      Verify selection.                                               */
